Move queues module into insights
