+++ static/app/views/insights/queues/charts/throughputChart.tsx
@@ -4,10 +4,10 @@ import {RateUnit} from 'sentry/utils/discover/fields';
 import {formatRate} from 'sentry/utils/formatters';
 import Chart, {ChartType} from 'sentry/views/insights/common/components/chart';
 import ChartPanel from 'sentry/views/insights/common/components/chartPanel';
-import {useProcessQueuesTimeSeriesQuery} from 'sentry/views/performance/queues/queries/useProcessQueuesTimeSeriesQuery';
-import {usePublishQueuesTimeSeriesQuery} from 'sentry/views/performance/queues/queries/usePublishQueuesTimeSeriesQuery';
-import type {Referrer} from 'sentry/views/performance/queues/referrers';
-import {CHART_HEIGHT} from 'sentry/views/performance/queues/settings';
+import {useProcessQueuesTimeSeriesQuery} from 'sentry/views/insights/queues/queries/useProcessQueuesTimeSeriesQuery';
+import {usePublishQueuesTimeSeriesQuery} from 'sentry/views/insights/queues/queries/usePublishQueuesTimeSeriesQuery';
+import type {Referrer} from 'sentry/views/insights/queues/referrers';
+import {CHART_HEIGHT} from 'sentry/views/insights/queues/settings';
 
 interface Props {
   referrer: Referrer;
