+++ static/app/views/insights/queues/queries/useQueuesByDestinationQuery.tsx
@@ -1,32 +1,40 @@
+import type {Sort} from 'sentry/utils/discover/fields';
+import {decodeScalar} from 'sentry/utils/queryString';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
+import {useLocation} from 'sentry/utils/useLocation';
 import {useSpanMetrics} from 'sentry/views/insights/common/queries/useDiscover';
-import type {Referrer} from 'sentry/views/performance/queues/referrers';
-import {DEFAULT_QUERY_FILTER} from 'sentry/views/performance/queues/settings';
+import {QueryParameterNames} from 'sentry/views/insights/common/views/queryParameters';
+import type {Referrer} from 'sentry/views/insights/queues/referrers';
+import {
+  DEFAULT_QUERY_FILTER,
+  TABLE_ROWS_LIMIT,
+} from 'sentry/views/insights/queues/settings';
 
 type Props = {
   referrer: Referrer;
   destination?: string;
   enabled?: boolean;
-  transaction?: string;
+  sort?: Sort;
 };
 
-export function useQueuesMetricsQuery({
-  destination,
-  transaction,
+export function useQueuesByDestinationQuery({
   enabled,
+  destination,
+  sort,
   referrer,
 }: Props) {
+  const location = useLocation();
+  const cursor = decodeScalar(location.query?.[QueryParameterNames.DESTINATIONS_CURSOR]);
+
   const mutableSearch = new MutableSearch(DEFAULT_QUERY_FILTER);
   if (destination) {
-    mutableSearch.addFilterValue('messaging.destination.name', destination);
-  }
-  if (transaction) {
-    mutableSearch.addFilterValue('transaction', transaction);
+    mutableSearch.addFilterValue('messaging.destination.name', destination, false);
   }
   const response = useSpanMetrics(
     {
       search: mutableSearch,
       fields: [
+        'messaging.destination.name',
         'count()',
         'count_op(queue.publish)',
         'count_op(queue.process)',
@@ -39,8 +47,9 @@ export function useQueuesMetricsQuery({
         'time_spent_percentage(app,span.duration)',
       ],
       enabled,
-      sorts: [],
-      limit: 10,
+      sorts: sort ? [sort] : [],
+      limit: TABLE_ROWS_LIMIT,
+      cursor,
     },
     referrer
   );
