+++ static/app/views/insights/queues/queuesTable.tsx
@@ -12,25 +12,31 @@ import type {CursorHandler} from 'sentry/components/pagination';
 import Pagination from 'sentry/components/pagination';
 import {t} from 'sentry/locale';
 import type {Organization} from 'sentry/types/organization';
+import {trackAnalytics} from 'sentry/utils/analytics';
 import {browserHistory} from 'sentry/utils/browserHistory';
 import type {EventsMetaType} from 'sentry/utils/discover/eventView';
 import {FIELD_FORMATTERS, getFieldRenderer} from 'sentry/utils/discover/fieldRenderers';
 import type {Sort} from 'sentry/utils/discover/fields';
-import {decodeScalar, decodeSorts} from 'sentry/utils/queryString';
-import useLocationQuery from 'sentry/utils/url/useLocationQuery';
+import {formatAbbreviatedNumber} from 'sentry/utils/formatters';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
 import {renderHeadCell} from 'sentry/views/insights/common/components/tableCells/renderHeadCell';
 import {QueryParameterNames} from 'sentry/views/insights/common/views/queryParameters';
-import {SpanFunction, type SpanMetricsResponse} from 'sentry/views/insights/types';
-import {useQueuesByTransactionQuery} from 'sentry/views/performance/queues/queries/useQueuesByTransactionQuery';
-import {Referrer} from 'sentry/views/performance/queues/referrers';
+import {useQueuesByDestinationQuery} from 'sentry/views/insights/queues/queries/useQueuesByDestinationQuery';
+import {Referrer} from 'sentry/views/insights/queues/referrers';
+import {
+  ModuleName,
+  SpanFunction,
+  SpanIndexedField,
+  type SpanMetricsResponse,
+} from 'sentry/views/insights/types';
 import {useModuleURL} from 'sentry/views/performance/utils/useModuleURL';
 
 type Row = Pick<
   SpanMetricsResponse,
   | 'sum(span.duration)'
-  | 'transaction'
+  | 'messaging.destination.name'
+  | 'avg(messaging.message.receive.latency)'
   | `avg_if(${string},${string},${string})`
   | `count_op(${string})`
 >;
@@ -39,13 +45,8 @@ type Column = GridColumnHeader<string>;
 
 const COLUMN_ORDER: Column[] = [
   {
-    key: 'transaction',
-    name: t('Transactions'),
-    width: COL_WIDTH_UNDEFINED,
-  },
-  {
-    key: 'span.op',
-    name: t('Type'),
+    key: 'messaging.destination.name',
+    name: t('Destination'),
     width: COL_WIDTH_UNDEFINED,
   },
   {
@@ -81,7 +82,7 @@ const COLUMN_ORDER: Column[] = [
 ];
 
 const SORTABLE_FIELDS = [
-  'transaction',
+  SpanIndexedField.MESSAGING_MESSAGE_DESTINATION_NAME,
   'count_op(queue.publish)',
   'count_op(queue.process)',
   'avg_if(span.duration,span.op,queue.process)',
@@ -94,46 +95,37 @@ type ValidSort = Sort & {
 };
 
 export function isAValidSort(sort: Sort): sort is ValidSort {
-  return (SORTABLE_FIELDS as unknown as string[]).includes(sort.field);
+  return (SORTABLE_FIELDS as ReadonlyArray<string>).includes(sort.field);
 }
 
-const DEFAULT_SORT = {
-  field: 'time_spent_percentage(app,span.duration)' as const,
-  kind: 'desc' as const,
-};
+interface Props {
+  sort: ValidSort;
+  destination?: string;
+  error?: Error | null;
+  meta?: EventsMetaType;
+}
 
-export function TransactionsTable() {
-  const organization = useOrganization();
+export function QueuesTable({error, destination, sort}: Props) {
   const location = useLocation();
+  const organization = useOrganization();
 
-  const locationQuery = useLocationQuery({
-    fields: {
-      destination: decodeScalar,
-      [QueryParameterNames.DESTINATIONS_SORT]: decodeScalar,
-    },
-  });
-  const sort =
-    decodeSorts(locationQuery[QueryParameterNames.DESTINATIONS_SORT])
-      .filter(isAValidSort)
-      .at(0) ?? DEFAULT_SORT;
-
-  const {data, isLoading, meta, pageLinks, error} = useQueuesByTransactionQuery({
-    destination: locationQuery.destination,
+  const {data, isLoading, meta, pageLinks} = useQueuesByDestinationQuery({
+    destination,
     sort,
-    referrer: Referrer.QUEUES_SUMMARY_TRANSACTIONS_TABLE,
+    referrer: Referrer.QUEUES_LANDING_DESTINATIONS_TABLE,
   });
 
   const handleCursor: CursorHandler = (newCursor, pathname, query) => {
     browserHistory.push({
       pathname,
-      query: {...query, [QueryParameterNames.TRANSACTIONS_CURSOR]: newCursor},
+      query: {...query, [QueryParameterNames.DESTINATIONS_CURSOR]: newCursor},
     });
   };
 
   return (
     <Fragment>
       <GridEditable
-        aria-label={t('Transactions')}
+        aria-label={t('Queues')}
         isLoading={isLoading}
         error={error}
         data={data}
@@ -157,7 +149,17 @@ export function TransactionsTable() {
         }}
       />
 
-      <Pagination pageLinks={pageLinks} onCursor={handleCursor} />
+      <Pagination
+        pageLinks={pageLinks}
+        onCursor={handleCursor}
+        paginationAnalyticsEvent={(direction: string) => {
+          trackAnalytics('insight.general.table_paginate', {
+            organization,
+            source: ModuleName.QUEUE,
+            direction,
+          });
+        }}
+      />
     </Fragment>
   );
 }
@@ -169,20 +171,8 @@ function renderBodyCell(
   location: Location,
   organization: Organization
 ) {
-  const op = row['span.op'];
-  const isProducer = op === 'queue.publish';
-  const isConsumer = op === 'queue.process';
   const key = column.key;
-  if (
-    row[key] === undefined ||
-    (isConsumer && ['count_op(queue.publish)'].includes(key)) ||
-    (isProducer &&
-      [
-        'count_op(queue.process)',
-        'avg(messaging.message.receive.latency)',
-        'avg_if(span.duration,span.op,queue.process)',
-      ].includes(key))
-  ) {
+  if (row[key] === undefined) {
     return (
       <AlignRight>
         <NoValue>{' \u2014 '}</NoValue>
@@ -190,8 +180,17 @@ function renderBodyCell(
     );
   }
 
-  if (key === 'transaction') {
-    return <TransactionCell transaction={row[key]} op={op} />;
+  if (key === 'messaging.destination.name' && row[key]) {
+    return <DestinationCell destination={row[key]} />;
+  }
+
+  if (key.startsWith('count')) {
+    return <AlignRight>{formatAbbreviatedNumber(row[key])}</AlignRight>;
+  }
+
+  if (key.startsWith('avg')) {
+    const renderer = FIELD_FORMATTERS.duration.renderFunc;
+    return renderer(key, row);
   }
 
   // Need to invert trace_status_rate(ok) to show error rate
@@ -208,22 +207,6 @@ function renderBodyCell(
     return row[column.key];
   }
 
-  if (key.startsWith('avg')) {
-    const renderer = FIELD_FORMATTERS.duration.renderFunc;
-    return renderer(key, row);
-  }
-
-  if (key === 'span.op') {
-    switch (row[key]) {
-      case 'queue.publish':
-        return t('Producer');
-      case 'queue.process':
-        return t('Consumer');
-      default:
-        return row[key];
-    }
-  }
-
   const renderer = getFieldRenderer(column.key, meta.fields, false);
   return renderer(row, {
     location,
@@ -232,18 +215,17 @@ function renderBodyCell(
   });
 }
 
-function TransactionCell({transaction, op}: {op: string; transaction: string}) {
+function DestinationCell({destination}: {destination: string}) {
   const moduleURL = useModuleURL('queue');
   const {query} = useLocation();
   const queryString = {
     ...query,
-    transaction,
-    'span.op': op,
+    destination,
   };
   return (
     <NoOverflow>
       <Link to={`${moduleURL}/destination/?${qs.stringify(queryString)}`}>
-        {transaction}
+        {destination}
       </Link>
     </NoOverflow>
   );
