+++ static/app/views/insights/queues/referrers.ts
