+++ static/app/views/insights/queues/settings.ts
