+++ static/app/views/performance/utils/useModuleURL.tsx
@@ -3,6 +3,7 @@ import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import {BASE_URL as CACHE_BASE_URL} from 'sentry/views/insights/cache/settings';
 import {BASE_URL as DB_BASE_URL} from 'sentry/views/insights/database/settings';
 import {BASE_URL as HTTP_BASE_URL} from 'sentry/views/insights/http/settings';
+import {BASE_URL as QUEUE_BASE_URL} from 'sentry/views/insights/queues/settings';
 import {ModuleName} from 'sentry/views/insights/types';
 import {BASE_URL as AI_BASE_URL} from 'sentry/views/llmMonitoring/settings';
 import {BASE_URL as RESOURCES_BASE_URL} from 'sentry/views/performance/browser/resources/settings';
@@ -10,7 +11,6 @@ import {BASE_URL as VITALS_BASE_URL} from 'sentry/views/performance/browser/webV
 import {BASE_URL as APP_STARTS_BASE_URL} from 'sentry/views/performance/mobile/appStarts/settings';
 import {BASE_URL as SCREEN_LOADS_BASE_URL} from 'sentry/views/performance/mobile/screenload/settings';
 import {BASE_URL as MOBILE_UI_BASE_URL} from 'sentry/views/performance/mobile/ui/settings';
-import {BASE_URL as QUEUE_BASE_URL} from 'sentry/views/performance/queues/settings';
 import {INSIGHTS_BASE_URL} from 'sentry/views/performance/settings';
 
 export const MODULE_BASE_URLS: Record<ModuleName, string> = {
